+++ src/ic-cdk/CHANGELOG.md
@@ -10,6 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change from pleco to tanton for the chess library in the chess example. (#345)
 
+## [0.6.8] - 2022-11-28
+
+### Added
+
+- Added composite queries via `#[query(composite = true)]`. (#344)
+
+  Composite queries cannot be run as update calls, but can make inter-canister calls to other query functions.
+
+- Implemented the canister timers API, located in module `ic_cdk::timer`. (#342)
+
 ## [0.6.7] - 2022-11-16
 
 ### Changed